feat:コードのリファクタリング
